requirements.txt
django==4.2.4
djangorestframework==3.14.0
django-filter==23.2
djangorestframework-api-keys==2.*
